requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tzdata>=2024.1
pyyaml>=6.0

# testing
pytest>=7.4.0
pytest-mock>=3.11.0

# lint & analysis
ruff>=0.11.0
vulture>=2.14
